// src/untils/enums.ts
export enum RoleType {
  CUSTOMER = 'Customer',
}

export enum PaymentStatusType {
  PENDING = 'Pending',
  PAID = 'Paid',
  FAILED = 'Failed',
}

export enum PaymentMethodType {
  ONLINE = 'Online',
  CASH = 'Cash',
}

export enum StatusType {
  PENDING = 'Pending',
  CONFIRMED = 'Confirmed',
  CANCELLED = 'Cancelled',
  COMPLETED = 'Completed',
}

export enum StatusTowType {
  SUCCEEDED = 'succeeded',
  FAILED = 'Failed',
  COMPLETED = 'Completed',
}

export enum RoomType {
  SINGLE = 'Single',
  DOUBLE = 'Double',
  SUITE = 'Suite',
  DELUXE = 'Deluxe',
}

export enum BedType {
  SINGLE = 'Single',
  DOUBLE = 'Double',
  QUEEN = 'Queen',
  KING = 'King',
}

export enum BathRoomType {
  PRIVATE = 'Private',
  SHARED = 'Shared',
}

export enum ViewType {
  SEA = 'Sea',
  CITY = 'City',
  GARDEN = 'Garden',
  POOL = 'Pool',
}

export enum SmokingPolicyType {
  ALLOWED = 'Allowed',
  NOTALLOWED = 'Not Allowed',
}

export enum ReservationStatusType {
  CONFIRMED = 'Confirmed',
  PENDING = 'Pending',
  CANCELLED = 'Cancelled',
}

export enum RoleTowType {
  OWNER = 'Owner',
  MANAGER = 'Manager',
  RECEPTIONIST = 'Receptionist',
  CLEANER = 'Cleaner',
}